refactor(new_split_chunks): split methods of `SplitChunksPlugin` into different files
